refactor(syscall): split task syscalls
